backend/pyproject.toml
[project]
name = "offline-clinic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [ "flask",
                 "flask-sqlalchemy",
                 "flask-migrate",
               "psycopg2-binary",]
